backend/requirements.txt
# FastAPI and server
fastapi>=0.121.2
uvicorn[standard]>=0.24.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.23  # Updated to 2.0.44 for Python 3.14 compatibility
alembic==1.12.1
psycopg2-binary==2.9.9
# asyncpg==0.29.0  # Note: Requires Visual C++ Build Tools for Python 3.14. Using psycopg instead.
psycopg[binary]>=3.2.0  # Async PostgreSQL driver with pre-built wheels for Python 3.14
greenlet>=3.0.0  # Required for psycopg async operations with SQLAlchemy

# Authentication & OAuth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
authlib==1.2.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# HTTP clients
httpx==0.25.1
requests==2.31.0

# Task queue
celery==5.3.4
redis==5.0.1

# AI & APIs
openai==1.3.5

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil>=2.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
aiosqlite==0.19.0

